backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
openai==1.58.1
stripe==11.4.1
razorpay==1.4.2
gspread==6.1.4
google-auth==2.37.0
shopifyapi==12.6.0
woocommerce==3.0.0
python-dotenv==1.0.1
redis==5.2.1
celery==5.4.0
jinja2==3.1.5
aiofiles==24.1.0
